rewrite some component

+++ core/ware.sh
@@ -7,24 +7,10 @@ fi
 
 password=$1 && shift
 
-if ! echo ${password} | sudo -kS echo "password is right" >/dev/null 2>&1; then
+if ! echo ${password} | env sudo -kS echo "password is right" >/dev/null 2>&1; then
     echo "incorrect password"
     exit
 fi
-
-
-path() {
-    realpath $(dirname $0)/..
-}
-
-tmp() {
-    ! [ -d $(path)/.tmp ] && mkdir $(path)/.tmp
-    echo $(path)/.tmp
-}
-
-output() {
-    echo "[warehouse]" "$@"
-}
 
 has() {
     if [ $# -eq 1 ]; then
@@ -48,13 +34,39 @@ sudo() {
     echo ${password} | env sudo -S "$@"
 }
 
+ware() {
+    if [[ $(os) == Ubuntu* ]]; then
+        sudo apt-get "$@"
+    fi
+    if [ $(os) = "macOS" ]; then
+        HOMEBREW_NO_AUTO_UPDATE=1 brew "$@"
+    fi
+}
+
+path() {
+    realpath $(dirname $0)/..
+}
+
+tmp() {
+    ! [ -d $(path)/.tmp ] && mkdir $(path)/.tmp
+    echo $(path)/.tmp
+}
+
+output() {
+    echo "[warehouse]" "$@"
+}
+
 numcmp() {
     awk -v n1="$1" -v n2="$2" 'BEGIN {printf (n1<n2?-1:(n1==n2?0:1))}'
 }
 
 sed() {
     input="${@: -1}"
-    env sed "$@" >${input}.tmp && mv ${input}.tmp ${input}
+    if [ $# -eq 2 ] && [ -e ${input} ]; then
+        env sed "$@" >${input}.tmp && mv ${input}.tmp ${input}
+    else
+        env sed "$@"
+    fi
 }
 
 tee() {
